moves error_codes to models_library
