+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core/_observer.py
@@ -6,7 +6,7 @@ from copy import deepcopy
 from math import floor
 
 from fastapi import FastAPI
-from servicelib.error_codes import create_error_code
+from models_library.error_codes import create_error_code
 
 from .....core.dynamic_services_settings.scheduler import (
     DynamicServicesSchedulerSettings,
