+++ services/web/server/src/simcore_service_webserver/login/_2fa_api.py
@@ -11,9 +11,9 @@ import asyncio
 import logging
 
 from aiohttp import web
+from models_library.error_codes import create_error_code
 from models_library.users import UserID
 from pydantic import BaseModel, Field
-from servicelib.error_codes import create_error_code
 from servicelib.logging_utils import LogExtra, get_log_record_extra, log_decorator
 from servicelib.utils_secrets import generate_passcode
 from settings_library.twilio import TwilioSettings
