+++ services/web/server/src/simcore_service_webserver/studies_dispatcher/_studies_access.py
@@ -20,10 +20,10 @@ from uuid import UUID, uuid5
 
 from aiohttp import web
 from aiohttp_session import get_session
+from models_library.error_codes import create_error_code
 from models_library.projects import ProjectID
 from servicelib.aiohttp import status
 from servicelib.aiohttp.typing_extension import Handler
-from servicelib.error_codes import create_error_code
 
 from .._constants import INDEX_RESOURCE_NAME
 from ..director_v2._core_computations import create_or_update_pipeline
